Implement TS to Reset component
